+++ polaris-tokens/src/themes/dark.ts
@@ -12,13 +12,13 @@ export const metaThemeDarkPartial = createMetaThemePartial({
     'color-icon-secondary': {value: colors.gray[12]},
     'color-text': {value: colors.gray[8]},
     'color-text-secondary': {value: colors.gray[11]},
-    'color-bg-surface-secondary-active': {value: colors.gray[13]},
-    'color-bg-surface-secondary-hover': {value: colors.gray[14]},
+    'color-bg-surface-secondary-active': {value: colors.whiteAlpha[9]},
+    'color-bg-surface-secondary-hover': {value: colors.whiteAlpha[6]},
     'color-bg-fill-transparent': {value: colors.whiteAlpha[8]},
     'color-bg-fill-brand': {value: colors.gray[1]},
     'color-text-brand-on-bg-fill': {value: colors.gray[15]},
     'color-bg-surface-hover': {value: colors.gray[14]},
-    'color-bg-fill-hover': {value: colors.gray[14]},
+    'color-bg-fill-hover': {value: 'rgba(37, 37, 37, 1)'},
     'color-bg-fill-transparent-hover': {value: colors.whiteAlpha[9]},
     'color-bg-fill-brand-hover': {value: colors.gray[5]},
     'color-bg-surface-selected': {value: colors.gray[13]},
@@ -29,8 +29,9 @@ export const metaThemeDarkPartial = createMetaThemePartial({
     'color-bg-fill-active': {value: colors.gray[13]},
     'color-bg-fill-transparent-active': {value: colors.whiteAlpha[10]},
     'color-bg-fill-brand-active': {value: colors.gray[4]},
+    'color-bg-fill-secondary-hover': {value: colors.whiteAlpha[8]},
     'color-bg-surface-brand-selected': {value: colors.gray[14]},
-    'color-border-secondary': {value: colors.gray[13]},
+    'color-border-secondary': {value: colors.gray[14]},
     'color-tooltip-tail-down-border-experimental': {
       value: 'rgba(60, 60, 60, 1)',
     },
@@ -38,16 +39,22 @@ export const metaThemeDarkPartial = createMetaThemePartial({
       value: 'rgba(71, 71, 71, 1)',
     },
     'color-border-gradient-experimental': {
-      value: `linear-gradient(to bottom, ${colors.whiteAlpha[9]}, ${colors.whiteAlpha[4]})`,
+      value: `linear-gradient(to bottom, ${colors.whiteAlpha[10]}, ${colors.whiteAlpha[4]})`,
     },
     'color-border-gradient-hover-experimental': {
-      value: `linear-gradient(to bottom, ${colors.whiteAlpha[9]}, ${colors.whiteAlpha[4]})`,
+      value: `linear-gradient(to bottom, ${colors.whiteAlpha[10]}, ${colors.whiteAlpha[4]})`,
     },
     'color-border-gradient-selected-experimental': {
-      value: `linear-gradient(to bottom, ${colors.blackAlpha[10]}, ${colors.whiteAlpha[10]})`,
+      value: `linear-gradient(to bottom, ${colors.blackAlpha[16]}, ${colors.blackAlpha[11]} 10%, ${colors.whiteAlpha[11]})`,
     },
     'color-border-gradient-active-experimental': {
-      value: `linear-gradient(to bottom, ${colors.whiteAlpha[10]}, ${colors.whiteAlpha[4]})`,
+      value: `linear-gradient(to bottom, ${colors.blackAlpha[16]}, ${colors.blackAlpha[11]} 10%, ${colors.whiteAlpha[11]})`,
+    },
+    'color-button-gradient-bg-fill-selected-experimental': {
+      value: `linear-gradient(to bottom, rgba(25, 25, 25, 1), rgba(19, 19, 19, 1) 50%, rgba(10, 10, 10, 1))`,
+    },
+    'color-button-gradient-bg-fill-active-experimental': {
+      value: `linear-gradient(to bottom, rgba(33, 33, 33, 1), rgba(30, 30, 30, 1) 50%, rgba(18, 18, 18, 1))`,
     },
   },
   shadow: {